pyproject.toml
[tool.poetry]
name = "pixel-expand"
version = "0.1.0"
description = ""
authors = ["Shuwiku <shuwiku@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
